added user roles

+++ resources/views/view-customer-account.blade.php
@@ -102,7 +102,9 @@
                             <th class="paymentDate">Payment Date</th>
                             <th class="payment">Payment</th>
                             <th class="paymentRecievedDate">Recieved Date</th>
-                            <th class="action">Action</th>
+                            @if (Auth::user()->role == 'admin')
+                                <th class="action">Action</th>
+                            @endif
                         </tr>
                     </thead>
                     <tbody>
@@ -119,12 +121,14 @@
                                 <td>{{ $customerPayment->payment_date }}</td>
                                 <td>{{ $customerAccount->currency . $customerPayment->payment }}</td>
                                 <td>{{ $customerPayment->payment_recieved_date }}</td>
-                                <td>
-                                    <a href="/customer-payment/edit/{{ $customerPayment->id }}"><button
-                                            class="primary">Edit</button></a>
-                                    <a href="/customer-payment/destroy/{{ $customerPayment->id }}"><button
-                                            class="danger">Delete</button></a>
-                                </td>
+                                @if (Auth::user()->role == 'admin')
+                                    <td>
+                                        <a href="/customer-payment/edit/{{ $customerPayment->id }}"><button
+                                                class="primary">Edit</button></a>
+                                        <a href="/customer-payment/destroy/{{ $customerPayment->id }}"><button
+                                                class="danger">Delete</button></a>
+                                    </td>
+                                @endif
                             </tr>
                         @endforeach
                     </tbody>
